GSF.Core: Modified TimeTagBase.cs to store seconds as a decimal rather than a double to prevent loss of precision when converting to and from TimeTag implementations.

+++ Source/Libraries/GSF.Core/UnixTimeTag.cs
@@ -76,7 +76,7 @@ namespace GSF
         /// Creates a new <see cref="UnixTimeTag"/>, given number of seconds since 1/1/1970.
         /// </summary>
         /// <param name="seconds">Number of seconds since 1/1/1970.</param>
-        public UnixTimeTag(double seconds)
+        public UnixTimeTag(decimal seconds)
             : base(BaseTicks, seconds)
         {
         }
@@ -86,7 +86,7 @@ namespace GSF
         /// </summary>
         /// <param name="seconds">Number of seconds since 1/1/1970.</param>
         public UnixTimeTag(uint seconds)
-            : base(BaseTicks, (double)seconds)
+            : base(BaseTicks, (decimal)seconds)
         {
         }
